requirements.txt
fastapi
pymongo
uvicorn
passlib
pydantic
python-jose[cryptography]
python-multipart
bcrypt
pyyaml
setuptools
starlette
httpx
pytest
motor